g85 - Move layout component folder
Description
Moved layout related templates into component folder.
