+++ imports/ui/layouts/HomeLayout.js
@@ -3,8 +3,8 @@ import '../../stylesheets/style.less';
 import './HomeLayout.html';
 
 //partials
-import '../partials/Footer.html';
-import '../partials/Header.html';
+import './component/Footer.html';
+import './component/Header.js';
 import '../partials/Overlay.js';
 
 TemplateController('HomeLayout',{
