+++ imports/ui/layouts/component/SideNav.js
@@ -1,7 +1,8 @@
 import { TemplateController } from 'meteor/space:template-controller';
 import './SideNav.html';
-import './NewMemoModal.js';
-import { Memos } from '../../api/memos.js';
+import '../../partials/NewMemoModal.js';
+import { Memos } from '../../../api/memos.js';
+
 TemplateController('SideNav',{
 	onCreated(){
 		const self = this;
